requirements.txt
torch==1.8.0
PyYAML==6.0
pymeshlab==2021.10
rich==10.16.1
scipy
